Linear Combiners for PSMs

PressureSensitiveMats.get_weights — Function

get_weights(comb, x, ref)

Returns the weights for a section of the PSM by the method comb

PressureSensitiveMats.combiner — Function

combiner(comb, x, ref)

Gets weights from combining method comb with a chosen reference sensor and applies it to x.

combiner(comb, x)

Gets weights from combining method comb and applies them to x

PressureSensitiveMats.snrmax — Function

snrmax(x)

Perform SNR-MAX combining on x. Equivalent to combiner(SNR_MAX(), x)

PressureSensitiveMats.mrc — Function

mrc(x, fs)

Perform MRC-PSD combining on x. Currently x must be less than 512 samples. Equivalent to combiner(MRC_PSD(fs), x)

PressureSensitiveMats.egc — Function

egc(x)

Perform equal gain combining on x. Equivalent to combiner(EGC(), x)

PressureSensitiveMats.pcc — Function

pcc(x)

Perform Pearson Correlation Coefficient combining on x. Equivalent to combiner(PCC2(), x)

PressureSensitiveMats.selection — Function

selection(x)

Perform selection combining on x. Equivalent to extract_ref(x)