Occupancy Detection for PSMs

Missing docstring.

Missing docstring for min_occupancy(x, m=300). Check Documenter's build log for details.

PressureSensitiveMats.breath_availability — Function

breath_availability(x)

Implements the breathing availbaility indicator from Long-Term Sleep Assessment by Unobtrusive Pressure Sensor Arrays - 2018, Soleimani.

Expects an array with dimensions n_mats x 8 x samples. Returns the maximum distance between mattress sensors and the mode of those distances.

PressureSensitiveMats.occupancy_detection — Function

occupancy_detection(x::AbstractVector; kwargs)

Implements signle sensor simple occupancy detection from Long-Term Sleep Assessment by Unobtrusive Pressure Sensor Arrays - 2018, Soleimani.

Arguments

β=350 : Average value of a sensor from an empty bed.

m=300 : Minimum number of samples to be declared an occupancy.

n=4 : Sensitivity parameter

occupancy_detection(x::AbstractMatrix; kwargs)

Implements simple occupancy detection from Long-Term Sleep Assessment by Unobtrusive Pressure Sensor Arrays - 2018, Soleimani.

Arguments

max_dist=false : Use breathing availability signal from the maximum distance method.

β=25000 : Average value of the sum of all sensors from an empty bed.

m=300 : Minimum number of samples to be declared an occupancy.

n=4 : Sensitivity parameter